backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
mysql-connector-python
pydantic
